requirements.txt
Flask==2.3.3
psutil==5.9.0
redis==4.6.0
psycopg2-binary==2.9.7
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.0.1